requirements.txt
alembic==0.8.10
appdirs==1.4.0
click==6.7
Flask==0.12
Flask-Caching==1.2.0
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
GeoAlchemy2==0.4.0
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==0.23
mercantile==0.9.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.6.2
pyparsing==2.1.10
python-editor==1.0.3
redis==2.10.5
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.5
Werkzeug==0.11.15
WTForms==2.1
